requirements.txt
# Requires Python 3.7 or later

# pip install -r requirements.txt


python-docx>=0.8.11
docx2pdf>=0.1.7
PyPDF2>=3.0.0

